refactor: move S3 utility functions to utils package
- Move generate_request_id() and generate_host_id() from s3_api.rs to utils/s3_utils.rs
- Move set_error_response() helper to utils/s3_utils.rs
- Update s3_api.rs to use utils::s3_utils::set_error_response()
- Add comprehensive documentation to utility functions
- Improve code organization by centralizing S3 utilities in utils package
- Better separation of concerns: s3_api.rs focuses on API logic, utils handles common functions
